migrating offer drawer to a component
