+++ frontend/src/components/task/task.js
@@ -21,7 +21,6 @@ import {
   Dialog,
   DialogContent,
   DialogTitle,
-  DialogActions,
   Grid,
   Typography,
   Button,
